requirements.txt
Django==1.8.4
django-appconf==1.0.1
django-compressor==1.5
django-cors-headers==1.1.0
django-extensions==1.5.9
django-modelcluster==1.0
django-sslify==0.2.7
django-storages==1.1.8
django-taggit==0.17.3
django-treebeard==3.0
django-uuidfield==0.5.0
djangocms-admin-style==0.2.8
djangorestframework==3.3.1
gunicorn==19.3.0
ordereddict==1.1
pytz==2015.7
six==1.10.0
urllib3==1.12
wsgiref==0.1.2
